+++ src/canvastools/ts/CanvasTools/Region/Polygon/TagsElement.ts
@@ -95,7 +95,7 @@ export class TagsElement extends TagsComponent {
             // Update primary tag text
             if (rebuildTags) {
                 this.primaryTagText.node.innerHTML = (this.tags.primary !== null) ? this.tags.primary.name : "";
-                this.textBox = this.primaryTagText.getBBox();
+                this.textBox = TagsComponent.getCachedBBox(this.primaryTagText);
             }
 
             const showTextLabel = (this.textBox.width + 10 <= this.width)
@@ -337,7 +337,7 @@ export class TagsElement extends TagsComponent {
 
         this.primaryTagText = paper.text(this.x, this.y, "");
         this.primaryTagText.addClass("primaryTagTextStyle");
-        this.textBox = this.primaryTagText.getBBox();
+        this.textBox = TagsComponent.getCachedBBox(this.primaryTagText);
 
         const pointsData = [];
         this.regionData.points.forEach((p) => {